avoid format mismatch warning
* test/testquorum.c (quorum_notification_fn): Add casts to
match formats.  Use unsigned formats for unsigned values.

git-svn-id: http://svn.fedorahosted.org/svn/corosync/trunk@1954 fd59a12c-fef9-0310-b244-a6a79926bd2f

+++ test/testquorum.c
@@ -20,9 +20,9 @@ static void quorum_notification_fn(
 	int i;
 
 	printf("quorum notification called \n");
-	printf("  quorate   = %d\n", quorate);
-	printf("  ring id   = %lld\n", ring_id);
-	printf("  num nodes = %d ", view_list_entries);
+	printf("  quorate   = %lu\n", (long unsigned int) quorate);
+	printf("  ring id   = %llu\n", (long long unsigned int) ring_id);
+	printf("  num nodes = %lu ", (long unsigned int) view_list_entries);
 
 	for (i=0; i<view_list_entries; i++) {
 		printf(" %d ", view_list[i]);
